requirements.txt
dns.resolver
whois
requests
ssl
socket
time
random
csv
tqdm
bs4
fuzzywuzzy